tests: organize system tests
