+++ .github/workflows/build.yaml
@@ -35,11 +35,8 @@ jobs:
       - name: Unit Tests
         run: go test ./... -v
 
-      - name: Integration Tests
-        run: go test ./... -tags integration -v
-
       - name: System Tests
-        run: go test ./tests -tags system -v
+        run: go test ./tests/... -tags system -v
 
   lint:
     runs-on: ubuntu-latest
