+++ tests/server/proxy_test.go
@@ -1,6 +1,6 @@
 //go:build system
 
-package tests
+package server
 
 import (
 	"bytes"
